BACKPORT: random: fix comment for unused random_min_urandom_seed
The variable random_min_urandom_seed is not needed any more as it
defined the reseeding behavior of the nonblocking pool. Though it is not
needed any more, it is left in the code for user space interface
compatibility.

Signed-off-by: Stephan Mueller <smueller@chronox.de>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
(cherry-picked from commit 5d0e5ea343a0f70351428476bcf8715e0731f26a)
Signed-off-by: Albert I <kras@raphielgang.org>