power: qpnp-linear-charger: Discharging status at full battery
Fix charger UI updation issue when the battery is at 100%
and the USB cable is removed. Move power_supply_changed call
outside the if check in usbin_valid_irq_handler and also
call power_supply_changed for batt_psy. This ensures that the
appropriate battery status uevent is sent out to take care
of case when the fastchg_irq is received before the usbin_valid_irq.

Change-Id: I92454a87a83a48c1e4baccd3f0773ac821b339bb
Signed-off-by: Shilpa Suresh <sbsure@codeaurora.org>