usb: gadget: f_ipc: Queue ZLP if data size is multiple of maxpacketsize
Currently the driver does not queue a zero length packet if the
data size is a multiple of maxpacketsize. If the host IPC driver
expects such a packet, then it is not completed immediately and
the IPC communication essentially breaks down because of size
mismatch.
Fix this by queueing ZLP if data packet size is a multiple of
wMacPacketSize.

Change-Id: Iecf8611d84089ac6e65ca8f864b9f538814e72d0
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>